+++ cmd/earth/main.go
@@ -530,7 +530,7 @@ func (app *earthApp) actionBuild(c *cli.Context) error {
 	var remoteConsoleAddr string
 	var err error
 	if app.interactiveDebugging {
-		debugServer := server.NewDebugServer(app.console)
+		debugServer := server.NewDebugServer(c.Context, app.console)
 		remoteConsoleAddr, err = debugServer.Start()
 		if err != nil {
 			app.console.Warnf("failed to open remote console listener: %v; interactive debugging disabled\n", err)
